fix memory issue

+++ labs/counter/Index.h
@@ -19,8 +19,8 @@ class Index {
     ~Index();
     // TODO: Implement hashFunction, insert, find, and remove methods
     void resizeAndRehash();
-    size_t hashFunction1(const std::string& key) const;
-    size_t hashFunction2(const std::string& key) const;
+    int hashFunction1(const std::string& key) const;
+    int hashFunction2(const std::string& key) const;
     void insert_index(const std::string& key, int value, List* list);
     List::Node* find(const std::string& key) const;
     List::Node* remove_index(const std::string& key, List* list);
